refactor(core): extract Blockchain trie ops to own module
Four impl Blockchain methods + supporting module-level items move out
of blockchain.rs into crates/sentrix-core/src/blockchain_trie_ops.rs:

  trie_root_at          — root hash at a specific version
  init_trie             — boot bind + AccountDB backfill on first init
  update_trie_for_block — per-block trie mutation, 280 LOC, carries
                          the 2026-05-07 extended-touch-list fork doc
  maybe_prune_trie      — every-5000-block background prune dispatch

Plus the module-level scaffolding for maybe_prune_trie:
  static PRUNE_RUNNING        — overlap guard (AtomicBool)
  pub(crate) trie_prune_disabled() — SENTRIX_DISABLE_TRIE_PRUNE reader

Largest single decomp this session at ~570 LOC moved. The remaining
impl Blockchain in blockchain.rs is now constructors / accessors /
fork delegators / consensus helpers — the trie + storage seams are
fully out.

Cleanup of blockchain.rs imports (now unused after the move):
  use hex
  use sentrix_primitives::transaction::{PROTOCOL_TREASURY, TOKEN_OP_ADDRESS}
  use sentrix_trie::address::{account_value_bytes, address_to_key}
  use std::sync::atomic::{AtomicBool, Ordering}

Tests in blockchain::tests that called trie_prune_disabled() now go
through crate::blockchain_trie_ops::trie_prune_disabled().

blockchain.rs drops from 3289 -> 2619 LOC (-670). 231 / 231 tests still
green; no behaviour change.
